backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
Flask-Cors==3.0.10
